+++ ChangeLog.md
@@ -4,6 +4,7 @@
 - Added configuration for interns for teleportation restriction on the revived player.
 - Fixed text typo in "Jetpack Thrusters" info.
 - Fixed issue with Jet Fuel configuration influencing Jetpack Thrusters in some aspects.
+- Fixed text typo in Sick beats description and added configuration to allow stamina buff apply on stamina drain overtime scenarios.
 
 </details>